Remove express_comment and fix email_confirmation
express_comment was incorrectly configured, and when corrected, it
turned out to be constant (1).  So I removed it.

The emailConfirmation element in the FCC json files is incorrectly
stored as a string "true" or "false" instead of the bare boolean
true or false.  I fixed standardize.boolean to accept the string form.